requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==1.4.52
python-dotenv==1.0.1
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.9.2
alembic==1.13.2